fix: delete record controller

+++ src/backend/server/controllers/records/records.controller.ts
@@ -2,7 +2,7 @@
 import { Record, RecordType, RecordTypeEnum } from '@prisma/client'
 import { prisma } from '../../prisma-client/prisma-client'
 import { getRecordTypeById } from '../record-types/record-types.controller'
-import { createRecordDetails } from '../record-details/record-details.controller'
+import { createRecordDetails, deleteRecordsDetailsByRecordId } from '../record-details/record-details.controller'
 import {
   createStock,
   getStockExisting,
@@ -45,7 +45,12 @@ const createRecord = async (
     const promiseArrayMovementsDetails = details.map(
       (stockIdAndQuantity: any) => {
         const { stockId, quantity } = stockIdAndQuantity
-        return createRecordDetails(stockId, recordCreated.id, quantity, subtotal)
+        return createRecordDetails(
+          stockId,
+          recordCreated.id,
+          quantity,
+          subtotal
+        )
       }
     )
     const allPromisesMovementsDetails = Promise.all(
@@ -104,6 +109,10 @@ const updateRecordById = async (
 
 const deleteRecordById = async (id: string) => {
   try {
+    if (!id) {
+      return
+    }
+    await deleteRecordsDetailsByRecordId(id)
     const deletedRecord: Record = await prisma.record.delete({
       where: { id }
     })
